Issue #81 test: Test Fake 수정
함수 추가 및 id Int -> Long

+++ app/src/test/java/com/drunkenboys/calendarun/data/schedule/local/FakeCalendarLocalDataSource.kt
@@ -14,7 +14,7 @@ class FakeCalendarLocalDataSource : CalendarLocalDataSource {
         TODO("Not yet implemented")
     }
 
-    override suspend fun fetchCalendar(id: Int): Calendar {
+    override suspend fun fetchCalendar(id: Long): Calendar {
         return Calendar(0, "test calendar", Date(), Date())
     }
 }
